Updated support-url list for limited unpack
Added list of files that SHOULD be unpacked from the utility packages.
This will be expanded in future to cover the standard MSYS etc, with the
corresponding DONT UNPACK tag.
Tags currently planned :
*+(file1 file2) - Unpack these colon delimited files.
*-(file1 file2) - Dont Unpack these colon delimited files.

Presence of one of these means the opposite case is wildcarded - ie if
there is a *+(), nothing else is unpacked except these files, and for
the *-() everything else IS unpacked.